Keep store unit test out of integration tests
The in-memory store source is shared by the library test utilities and
integration-test helpers. Keep its unit test in the library-only module
so each integration binary does not register and rerun it.
