+++ src/io/in_memory_store.rs
@@ -222,29 +222,3 @@ impl MigratableKVStore for InMemoryStore {
 
 unsafe impl Sync for InMemoryStore {}
 unsafe impl Send for InMemoryStore {}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[tokio::test]
-	async fn in_memory_store_list_all_keys() {
-		let store = InMemoryStore::new();
-
-		KVStore::write(&store, "ns_a", "sub_a", "key_a", vec![1u8]).await.unwrap();
-		KVStore::write(&store, "ns_a", "sub_b", "key_b", vec![2u8]).await.unwrap();
-		KVStore::write(&store, "ns_b", "", "key_c", vec![3u8]).await.unwrap();
-
-		let mut keys = MigratableKVStore::list_all_keys(&store).await.unwrap();
-		keys.sort();
-
-		assert_eq!(
-			keys,
-			vec![
-				("ns_a".to_string(), "sub_a".to_string(), "key_a".to_string()),
-				("ns_a".to_string(), "sub_b".to_string(), "key_b".to_string()),
-				("ns_b".to_string(), "".to_string(), "key_c".to_string()),
-			]
-		);
-	}
-}
